fix(pdf-render): fixed logic in task models
